added swal for courses

+++ resources/js/components/Courses.vue
@@ -47,6 +47,7 @@
 
 <script>
 import { mapGetters } from 'vuex';
+import Swal from 'sweetalert2';
 
 export default{
     name: "courses",
@@ -61,7 +62,38 @@ export default{
         },
     methods: {
         deleteCourse(id) {
-            this.$store.dispatch("removeCourse", id)
+            this.$swal.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Yes, delete it!'
+            }).then((response)=>{
+                if (response.isConfirmed){
+                    this.$store.dispatch("removeCourse", id)
+                    this.$swal.fire(
+                        {
+                            title: 'Deleted!',
+                            text: 'Your file has been deleted.',
+                            icon: 'success',
+                            showConfirmButton: false,
+                            timer: 1500
+                        }
+                    )
+                } else if (
+                    response.dismiss === Swal.DismissReason.cancel
+                ) {
+                    this.$swal.fire({
+                        title: 'Canceled',
+                        text: 'Your imaginary file is safe :)',
+                        icon: 'error',
+                        showConfirmButton: false,
+                        timer: 1500
+                    })
+                }
+            })
         },
     }, 
 }
